requirements.txt
fastapi
uvicorn
openai
requests
pydantic
python-dotenv
numpy
faiss-cpu
